New folder/apps/api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
pyswisseph==2.10.3.2
pytz==2024.1
timezonefinder==6.5.0
geopy==2.4.1
python-dateutil==2.8.2
pydantic[email]==2.5.3
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
APScheduler==3.10.4
PyYAML==6.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Dev tools
ruff==0.1.14
black==24.1.1
mypy==1.8.0
